requirements.txt
django
django-suit
django-taggit
